requirements.txt
langchain
langchain-google-genai
langchain-community
python-dotenv
pypdf
chromadb
streamlit
faiss-cpu
